refactoring into crud
